Modified library allincludes header and moved into 'utilities' subdir.
